Don't store callback_info
This also eliminates a useless check in gig_callback_to_c.
The GIInfoType of a callback should always be GI_INFO_TYPE_CALLBACK.
If we wanted to figure out the actual callback type, we should
use g_registered_type_*, but it's not guaranteed, that callbacks infos have
GTypes (sadly).

For now, we forgo any type checking here. We may should revisit this decision later.