Register hashtables correctly.
Add an association between G_TYPE_HASH_TABLE and (@ (oop goops) <hashtable>)
at a type level, rather than rewriting the type_specializer in gig_function.
The latter needs a full GigTypeMeta to resolve ambiguous cases, such as pointer
and character types, the former is based on a hash lookup using the GType as
key.  As a result, this works even without GigTypeMeta, such as when dealing
with the GValue API.