Hardcode extension paths.
This is the recipe, that finally makes guile-gi work as intended in Guix and
fixes #72.
It turns out, that we still get compilation errors if we just hardcode the
paths.  Therefore, we have to first build and install the library, then
hardcode the paths, then build it again, check and install.  This procedure
is safe -- albeit time-consuming -- since Guix will only really "install"
the package if all steps succeed.